Actually fix the issue, not the warning.

+++ src/noggit/blp_texture.cpp
@@ -128,8 +128,8 @@ namespace noggit
                                               )
       );
 
-    int width (header->resx);
-    int height (header->resy);
+    size_t width (header->resx);
+    size_t height (header->resy);
 
     unsigned int* result_buffer (new unsigned int[width * height]);
 
@@ -156,9 +156,9 @@ namespace noggit
         int alpha_bit_count (0);
         const unsigned char* alpha_pointer (color_pointer + width * height);
 
-        for (size_t y (0); y < (size_t)height; ++y)
+        for (size_t y (0); y < height; ++y)
         {
-          for (size_t x (0); x < (size_t)width; ++x)
+          for (size_t x (0); x < width; ++x)
           {
             unsigned int color_value (palette[*color_pointer++]);
 
@@ -190,9 +190,9 @@ namespace noggit
       }
       else
       {
-        for (size_t y (0); y < (size_t)height; ++y)
+        for (size_t y (0); y < height; ++y)
         {
-          for (size_t x (0); x < (size_t)width; ++x)
+          for (size_t x (0); x < width; ++x)
           {
             unsigned int color_value (palette[*color_pointer++]);
             *result_pointer++ = (color_value & 0x00FF0000) >> 16
@@ -248,8 +248,8 @@ namespace noggit
                        : alphatypes[alpha_type]
                        );
 
-    int width (header->resx);
-    int height (header->resy);
+    size_t width (header->resx);
+    size_t height (header->resy);
 
     for (size_t mipmap_level (0); mipmap_level < 16; ++mipmap_level)
     {
